requirements.txt
pandas
numpy
matplotlib
seaborn
scipy
scikit-learn
xgboost
statsmodels
imbalanced-learn
pyyaml
mlflow